backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0   # JWT
httpx==0.27.0                       # Google OAuth HTTP calls
python-dotenv==1.0.1
google-generativeai==0.5.4          # Gemini SDK
alembic==1.13.1                     # DB migrations (optional but good practice)